chat - refactor status entry
